Create Fake Backend
Create a folder named _helpers in the /src/app folder. 
The fake backend contains a handleRoute function that checks if the request matches one of the faked routes in the switch statement.  Requests to the authenticate route are handled by the authenticate function which checks the username and password against an array of hardcoded users. If the username and password are correct then an ok response is returned with the user details and a fake jwt token, otherwise an error response is returned.