Update Nav in App Component Template
The updated nav contains just two links - Home and Logout. The home link navigates to the home route ("/") using the routerLink attribute directive. The logout link calls the logout() method on click by using the (click)="logout()" event binding attribute.

The *ngIf="currentUser" directive renders the nav if the currentUser property contains a value, which indicates that the user is logged in.